Update pathfinding logic and exception handling

+++ Map/Navigation/NavigationService.cs
@@ -15,17 +15,17 @@ namespace SBad.Map.Navigation
 
 		public int MaxRow { get; set; }
 		public int MaxCol { get; set; }
-		public int[,] Weight;
+		public decimal[,] Weight;
 		public IPathfinder Pathfinder { get; set; }
-        public FloorPlan FloorPlan { get; set; }
+		public FloorPlan FloorPlan { get; set; }
 
 		public void Refresh(FloorPlan floorPlan)
 		{
-            FloorPlan = floorPlan;
+			FloorPlan = floorPlan;
 			MaxRow = floorPlan.FloorTiles.Max(x => x.Y);
 			MaxCol = floorPlan.FloorTiles.Max(x => x.X);
 
-			Weight = new int[MaxCol + 1, MaxRow + 1];
+			Weight = new decimal[MaxCol + 1, MaxRow + 1];
 
 			for (int row = 0; row <= MaxRow; row++)
 			{
@@ -37,31 +37,31 @@ namespace SBad.Map.Navigation
 			}
 		}
 
-        public List<Point> Navigate(Point start, Point end, List<Point> path)
-        {
-            List<Point> pathSoFar = path;
-            FloorRoom room = FloorPlan.GetRoom(start);
+		public List<Point> Navigate(Point start, Point end, List<Point> path)
+		{
+			List<Point> pathSoFar = path;
+			FloorRoom room = FloorPlan.GetRoom(start);
 
-            if (room == FloorPlan.GetRoom((end)))
-            {
-                pathSoFar.AddRange(FindPath(start, end));
-                return pathSoFar;
-            }
-            else
-            {
-                var doors = room.DoorTiles;
-                foreach(var door in room.DoorTiles)
-                {
-                    pathSoFar.AddRange(Pathfinder.FindPath(start, door.Point));
-                    var exitTile = door.ExitTile;
+			if (room == FloorPlan.GetRoom((end)))
+			{
+				pathSoFar.AddRange(FindPath(start, end));
+				return pathSoFar;
+			}
+			else
+			{
+				var doors = room.DoorTiles;
+				foreach (var door in room.DoorTiles)
+				{
+					pathSoFar.AddRange(Pathfinder.FindPath(start, door.Point));
+					var exitTile = door.ExitTile;
 
-                    Navigate(exitTile.Point, end, pathSoFar);
-                }
-            }
+					Navigate(exitTile.Point, end, pathSoFar);
+				}
+			}
 
 
-            return pathSoFar;
-        }
+			return pathSoFar;
+		}
 
 		public List<Point> FindPath(Point start, Point end)
 		{
